# internal/presentation/web/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "A comprehensive cryptocurrency payment processing API for merchants and customers.",
        "title": "Crypto Checkout API",
        "termsOfService": "https://crypto-checkout.com/terms",
        "contact": {
            "name": "API Support",
            "url": "https://crypto-checkout.com/support",
            "email": "support@crypto-checkout.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/v1/invoices": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Create a new invoice for cryptocurrency payment processing",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Invoices"
                ],
                "summary": "Create a new invoice",
                "parameters": [
                    {
                        "description": "Invoice creation request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/web.CreateInvoiceRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Invoice created successfully",
                        "schema": {
                            "$ref": "#/definitions/web.CreateInvoiceResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request parameters",
                        "schema": {
                            "$ref": "#/definitions/web.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized - Invalid API key",
                        "schema": {
                            "$ref": "#/definitions/web.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/web.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/invoices/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieve detailed information about a specific invoice",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Invoices"
                ],
                "summary": "Get invoice details",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Invoice ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Invoice details retrieved successfully",
                        "schema": {
                            "$ref": "#/definitions/web.CreateInvoiceResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid invoice ID",
                        "schema": {
                            "$ref": "#/definitions/web.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized - Invalid API key",
                        "schema": {
                            "$ref": "#/definitions/web.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Invoice not found",
                        "schema": {
                            "$ref": "#/definitions/web.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/web.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/invoice/{id}/qr": {
            "get": {
                "description": "Generate and return a QR code image for invoice payment",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "image/png"
                ],
                "tags": [
                    "Invoices"
                ],
                "summary": "Get invoice QR code",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Invoice ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "QR code image",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Invalid invoice ID",
                        "schema": {
                            "$ref": "#/definitions/web.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Invoice not found",
                        "schema": {
                            "$ref": "#/definitions/web.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/web.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "web.CreateInvoiceRequest": {
            "type": "object",
            "required": [
                "items",
                "tax_rate"
            ],
            "properties": {
                "items": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "$ref": "#/definitions/web.InvoiceItemRequest"
                    }
                },
                "tax_rate": {
                    "type": "string"
                }
            }
        },
        "web.CreateInvoiceResponse": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "customer_url": {
                    "type": "string"
                },
                "expires_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "invoice_url": {
                    "type": "string"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/web.InvoiceItemResponse"
                    }
                },
                "payment_address": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "subtotal": {
                    "type": "string"
                },
                "tax_amount": {
                    "type": "string"
                },
                "tax_rate": {
                    "type": "string"
                },
                "total": {
                    "type": "string"
                },
                "usdt_amount": {
                    "description": "API.md required fields",
                    "type": "string"
                }
            }
        },
        "web.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "details": {
                    "type": "object",
                    "additionalProperties": true
                },
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "request_id": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
        "web.InvoiceItemRequest": {
            "type": "object",
            "required": [
                "description",
                "quantity",
                "unit_price"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "quantity": {
                    "type": "string"
                },
                "unit_price": {
                    "type": "string"
                }
            }
        },
        "web.InvoiceItemResponse": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "quantity": {
                    "type": "string"
                },
                "total": {
                    "type": "string"
                },
                "unit_price": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "API Key authentication. Use format: Bearer sk_live_xxx or Bearer sk_test_xxx",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "BearerAuth": {
            "description": "JWT Bearer token authentication. Use format: Bearer \u003cjwt_token\u003e",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}